pyproject.toml
[project]
name = "litestar-htmx-tailwind"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "litestar[jinja,standard,cryptography]>=2.0.0",
    "uvicorn[standard]>=0.27.0",
    "aiosqlite>=0.20.0",
    "yoyo-migrations>=8.2.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "aiosmtplib>=3.0.0",
    "click>=8.0.0"
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-xdist>=3.5.0",
    "httpx>=0.26.0",
    "ty>=0.0.19",
    "ruff>=0.15.4",
]

[tool.ty.environment]
root = ["./src"]
